backend/requirements.txt
django
djangorestframework
torch
torchvision
timm
opencv-python
numpy
pillow
matplotlib
kagglehub
pytorch-grad-cam